+++ adapters/MoPub/MoPubAdapter/GADMAdapterMoPub.m
@@ -1,6 +1,9 @@
 #import "GADMAdapterMoPub.h"
+
+#import "GADMAdapterMoPubConstants.h"
 #import "GADMAdapterMoPubSingleton.h"
 #import "GADMAdapterMoPubUtils.h"
+#import "GADMAdapterMopubUnifiedNativeAd.h"
 #import "GADMoPubNetworkExtras.h"
 #import "MPAdView.h"
 #import "MPImageDownloadQueue.h"
@@ -16,10 +19,8 @@
 #import "MPStaticNativeAdRenderer.h"
 #import "MPStaticNativeAdRendererSettings.h"
 #import "MoPub.h"
-#import "MoPubAdapterConstants.h"
-#import "MoPubAdapterMediatedNativeAd.h"
 
-static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelegates;
+static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMAdapterMoPubInterstitialDelegates;
 
 @interface GADMAdapterMoPub () <MPNativeAdDelegate,
                                 MPAdViewDelegate,
@@ -30,28 +31,35 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
   /// Connector from Google Mobile Ads SDK to receive ad configurations.
   __weak id<GADMAdNetworkConnector> _connector;
 
+  /// Array of ad loader options.
   NSArray<GADAdLoaderOptions *> *_nativeAdOptions;
 
+  /// MoPub banner ad.
   MPAdView *_bannerAd;
 
+  /// MoPub interstitial ad.
   MPInterstitialAdController *_interstitialAd;
 
+  /// MoPub native ad.
   MPNativeAd *_nativeAd;
 
-  MoPubAdapterMediatedNativeAd *_mediatedAd;
+  /// MoPub native ad wrapper.
+  GADMAdapterMopubUnifiedNativeAd *_mediatedAd;
 
+  /// MoPub's image download queue.
   MPImageDownloadQueue *_imageDownloadQueue;
 
-  NSMutableDictionary<NSString *, GADNativeAdImage *> *_imagesDictionary;
-
+  /// Ad loader options for configuring the view of native ads.
   GADNativeAdViewAdOptions *_nativeAdViewAdOptions;
 
-  BOOL _shouldDownloadImages;
+  /// Serializes GADMAdapterMoPubInterstitialDelegates usage.
+  dispatch_queue_t _lockQueue;
 }
 
 + (void)load {
-  GADMInterstitialAdapterDelegates = [NSMapTable mapTableWithKeyOptions:NSPointerFunctionsStrongMemory
-                                                       valueOptions:NSPointerFunctionsWeakMemory];
+  GADMAdapterMoPubInterstitialDelegates =
+      [NSMapTable mapTableWithKeyOptions:NSPointerFunctionsStrongMemory
+                            valueOptions:NSPointerFunctionsWeakMemory];
 }
 
 + (NSString *)adapterVersion {
@@ -67,6 +75,7 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
   if (self) {
     _connector = connector;
     _imageDownloadQueue = [[MPImageDownloadQueue alloc] init];
+    _lockQueue = dispatch_queue_create("mopub-interstitialAdapterDelegates", DISPATCH_QUEUE_SERIAL);
   }
   return self;
 }
@@ -130,29 +139,26 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
   id<GADMAdNetworkConnector> strongConnector = _connector;
   NSString *publisherID = strongConnector.credentials[kGADMAdapterMoPubPubIdKey];
 
-  @synchronized(GADMInterstitialAdapterDelegates) {
-    if ([GADMInterstitialAdapterDelegates objectForKey:publisherID]) {
-      NSError *adapterError = [NSError
-          errorWithDomain:kGADMAdapterMoPubErrorDomain
-                     code:kGADErrorInvalidRequest
-                 userInfo:@{
-                   NSLocalizedDescriptionKey : @"Unable to request a second ad using the sample "
-                                               @"publisher ID while the first ad is still active."
-                 }];
+  dispatch_async(_lockQueue, ^{
+    if ([GADMAdapterMoPubInterstitialDelegates objectForKey:publisherID]) {
+      NSError *adapterError = GADMAdapterMoPubErrorWithCodeAndDescription(
+          kGADErrorInvalidRequest, @"Unable to request a second ad using the same publisher ID "
+                                   @"while the first ad is still active.");
       [strongConnector adapter:self didFailAd:adapterError];
       return;
     } else {
-      GADMAdapterMoPubMapTableSetObjectForKey(GADMInterstitialAdapterDelegates, publisherID, self);
+      GADMAdapterMoPubMapTableSetObjectForKey(GADMAdapterMoPubInterstitialDelegates, publisherID,
+                                              self);
     }
-  }
+  });
 
   CLLocation *currentlocation = [[CLLocation alloc] initWithLatitude:strongConnector.userLatitude
                                                            longitude:strongConnector.userLongitude];
 
   _interstitialAd = [MPInterstitialAdController interstitialAdControllerForAdUnitId:publisherID];
   _interstitialAd.delegate = self;
-  _interstitialAd.keywords = [self getKeywords:false];
-  _interstitialAd.userDataKeywords = [self getKeywords:true];
+  _interstitialAd.keywords = [self getKeywords:NO];
+  _interstitialAd.userDataKeywords = [self getKeywords:YES];
   _interstitialAd.location = currentlocation;
 
   MPLogDebug(@"Requesting Interstitial Ad from MoPub Ad Network.");
@@ -175,12 +181,12 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
 }
 
 - (void)interstitialDidFailToLoadAd:(MPInterstitialAdController *)interstitial {
-  NSError *adapterError = [NSError errorWithDomain:kGADMAdapterMoPubErrorDomain
-                                              code:kGADErrorMediationNoFill
-                                          userInfo:nil];
-  @synchronized(GADMInterstitialAdapterDelegates) {
-    GADMAdapterMoPubMapTableRemoveObjectForKey(GADMInterstitialAdapterDelegates, interstitial.adUnitId);
-  }
+  NSError *adapterError =
+      GADMAdapterMoPubErrorWithCodeAndDescription(kGADErrorNoFill, @"Mopub failed to fill the ad.");
+  dispatch_async(_lockQueue, ^{
+    GADMAdapterMoPubMapTableRemoveObjectForKey(GADMAdapterMoPubInterstitialDelegates,
+                                               interstitial.adUnitId);
+  });
   [_connector adapter:self didFailAd:adapterError];
 }
 
@@ -193,9 +199,10 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
 }
 
 - (void)interstitialDidDisappear:(MPInterstitialAdController *)interstitial {
-  @synchronized(GADMInterstitialAdapterDelegates) {
-    GADMAdapterMoPubMapTableRemoveObjectForKey(GADMInterstitialAdapterDelegates, interstitial.adUnitId);
-  }
+  dispatch_async(_lockQueue, ^{
+    GADMAdapterMoPubMapTableRemoveObjectForKey(GADMAdapterMoPubInterstitialDelegates,
+                                               interstitial.adUnitId);
+  });
   [_connector adapterDidDismissInterstitial:self];
 }
 
@@ -214,8 +221,8 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
 
   _bannerAd = [[MPAdView alloc] initWithAdUnitId:publisherID];
   _bannerAd.delegate = self;
-  _bannerAd.keywords = [self getKeywords:false];
-  _bannerAd.userDataKeywords = [self getKeywords:true];
+  _bannerAd.keywords = [self getKeywords:NO];
+  _bannerAd.userDataKeywords = [self getKeywords:YES];
   _bannerAd.location = currentlocation;
 
   MPLogDebug(@"Requesting Banner Ad from MoPub Ad Network.");
@@ -233,14 +240,9 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
 }
 
 - (void)adViewDidFailToLoadAd:(MPAdView *)view {
-  NSString *errorDescription = [NSString stringWithFormat:@"Mopub failed to fill the ad."];
-  NSDictionary *errorInfo =
-      [NSDictionary dictionaryWithObjectsAndKeys:errorDescription, NSLocalizedDescriptionKey, nil];
-
-  [_connector adapter:self
-            didFailAd:[NSError errorWithDomain:kGADErrorDomain
-                                          code:kGADErrorInvalidRequest
-                                      userInfo:errorInfo]];
+  NSError *error = GADMAdapterMoPubErrorWithCodeAndDescription(kGADErrorInvalidRequest,
+                                                               @"Mopub failed to fill the ad.");
+  [_connector adapter:self didFailAd:error];
 }
 
 - (void)willLeaveApplicationFromAd:(MPAdView *)view {
@@ -268,6 +270,8 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
 - (void)getNativeAdWithAdTypes:(NSArray<GADAdLoaderAdType> *)adTypes
                        options:(NSArray<GADAdLoaderOptions *> *)options {
   id<GADMAdNetworkConnector> strongConnector = _connector;
+  _nativeAdOptions = options;
+
   MPStaticNativeAdRendererSettings *settings = [[MPStaticNativeAdRendererSettings alloc] init];
   MPNativeAdRendererConfiguration *config =
       [MPStaticNativeAdRenderer rendererConfigurationWithRendererSettings:settings];
@@ -277,17 +281,15 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
                                                          rendererConfigurations:@[ config ]];
 
   MPNativeAdRequestTargeting *targeting = [MPNativeAdRequestTargeting targeting];
-  targeting.keywords = [self getKeywords:false];
-  targeting.userDataKeywords = [self getKeywords:true];
+  targeting.keywords = [self getKeywords:NO];
+  targeting.userDataKeywords = [self getKeywords:YES];
   CLLocation *currentlocation = [[CLLocation alloc] initWithLatitude:strongConnector.userLatitude
                                                            longitude:strongConnector.userLongitude];
   targeting.location = currentlocation;
-  NSSet *desiredAssets = [NSSet
+  NSSet<NSString *> *desiredAssets = [NSSet
       setWithObjects:kAdTitleKey, kAdTextKey, kAdIconImageKey, kAdMainImageKey, kAdCTATextKey, nil];
   targeting.desiredAssets = desiredAssets;
-
   adRequest.targeting = targeting;
-  _nativeAdOptions = options;
 
   [[GADMAdapterMoPubSingleton sharedInstance] initializeMoPubSDKWithAdUnitID:publisherID
                                                            completionHandler:^{
@@ -299,175 +301,138 @@ static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMInterstitialAdapterDelega
   MPLogDebug(@"Requesting Native Ad from MoPub Ad Network.");
   [adRequest startWithCompletionHandler:^(MPNativeAdRequest *request, MPNativeAd *response,
                                           NSError *error) {
-    [self handleNativeAdOptions:request
-                   withResponse:response
-                      withError:error
-                    withOptions:self->_nativeAdOptions];
+    [self handleNativeAdWithResponse:response withError:error];
   }];
 }
 
-- (void)handleNativeAdOptions:(MPNativeAdRequest *)request
-                 withResponse:(MPNativeAd *)response
-                    withError:(NSError *)error
-                  withOptions:(NSArray<GADAdLoaderOptions *> *)options {
+- (void)handleNativeAdWithResponse:(MPNativeAd *)response withError:(NSError *)error {
   if (error) {
     [_connector adapter:self didFailAd:error];
-  } else {
-    _nativeAd = response;
-    _nativeAd.delegate = self;
-    _shouldDownloadImages = YES;
-
-    if (options != nil) {
-      for (GADAdLoaderOptions *loaderOptions in options) {
-        if ([loaderOptions isKindOfClass:[GADNativeAdImageAdLoaderOptions class]]) {
-          GADNativeAdImageAdLoaderOptions *imageOptions =
-              (GADNativeAdImageAdLoaderOptions *)loaderOptions;
-          _shouldDownloadImages = !imageOptions.disableImageLoading;
-        } else if ([loaderOptions isKindOfClass:[GADNativeAdViewAdOptions class]]) {
-          _nativeAdViewAdOptions = (GADNativeAdViewAdOptions *)loaderOptions;
-        }
-      }
-    }
-    [self loadNativeAdImages];
+    return;
   }
+
+  _nativeAd = response;
+  _nativeAd.delegate = self;
+  GADMAdapterMoPub *__weak weakSelf = self;
+  [self preCacheNativeImagesWithCompletionHandler:^(
+            NSDictionary<NSString *, GADNativeAdImage *> *_Nullable imagesDictionary) {
+    GADMAdapterMoPub *strongSelf = weakSelf;
+    if (!strongSelf) {
+      return;
+    }
+    id<GADMAdNetworkConnector> strongConnector = strongSelf->_connector;
+    if (!imagesDictionary[kAdIconImageKey] || !imagesDictionary[kAdMainImageKey]) {
+      NSError *adapterError = GADMAdapterMoPubErrorWithCodeAndDescription(
+          kGADErrorNetworkError, @"Failed to download images.");
+      [strongConnector adapter:strongSelf didFailAd:adapterError];
+      return;
+    }
+    strongSelf->_mediatedAd = [[GADMAdapterMopubUnifiedNativeAd alloc]
+        initWithMoPubNativeAd:strongSelf->_nativeAd
+                    mainImage:imagesDictionary[kAdMainImageKey]
+                    iconImage:imagesDictionary[kAdIconImageKey]
+          nativeAdViewOptions:strongSelf->_nativeAdViewAdOptions
+                networkExtras:strongConnector.networkExtras];
+    [strongConnector adapter:strongSelf didReceiveMediatedUnifiedNativeAd:strongSelf->_mediatedAd];
+  }];
 }
 
 #pragma mark - Helper methods for downloading images
 
-- (void)loadNativeAdImages {
+- (nullable NSDictionary<NSString *, NSURL *> *)imageURLs {
   id<GADMAdNetworkConnector> strongConnector = _connector;
-  NSMutableArray<NSURL *> *imageURLs = [NSMutableArray array];
-  NSError *adapterError = [NSError errorWithDomain:kGADMAdapterMoPubErrorDomain
-                                              code:kGADErrorReceivedInvalidResponse
-                                          userInfo:nil];
+  NSMutableDictionary<NSString *, NSURL *> *imageURLDictionary = [[NSMutableDictionary alloc] init];
 
-  for (NSString *key in [_nativeAd.properties allKeys]) {
-    if ([[key lowercaseString] hasSuffix:@"image"] &&
-        [[_nativeAd.properties objectForKey:key] isKindOfClass:[NSString class]]) {
-      if ([_nativeAd.properties objectForKey:key]) {
-        NSURL *URL = [NSURL URLWithString:_nativeAd.properties[key]];
-        if (URL != nil) {
-          GADMAdapterMoPubMutableArrayAddObject(imageURLs, URL);
-        } else {
-          [strongConnector adapter:self didFailAd:adapterError];
-          return;
-        }
-      } else {
-        [strongConnector adapter:self didFailAd:adapterError];
-        return;
+  for (NSString *imageKey in [_nativeAd.properties allKeys]) {
+    if ([imageKey.lowercaseString hasSuffix:@"image"] &&
+        [_nativeAd.properties[imageKey] isKindOfClass:[NSString class]]) {
+      NSString *nativeAdImageURLString = _nativeAd.properties[imageKey];
+      NSURL *imageURL = [NSURL URLWithString:nativeAdImageURLString];
+      if (!imageURL) {
+        NSString *errorString =
+            [NSString stringWithFormat:@"Invalid URL: %@", nativeAdImageURLString];
+        NSError *error = GADMAdapterMoPubErrorWithCodeAndDescription(kGADErrorMediationAdapterError,
+                                                                     errorString);
+        [strongConnector adapter:self didFailAd:error];
+        return nil;
       }
+      GADMAdapterMoPubMutableDictionarySetObjectForKey(imageURLDictionary, imageKey, imageURL);
     }
   }
-  [self precacheImagesWithURL:imageURLs];
+
+  return imageURLDictionary;
 }
 
-- (NSString *)returnImageKey:(NSString *)imageURL {
-  for (NSString *key in [_nativeAd.properties allKeys]) {
-    if ([[key lowercaseString] hasSuffix:@"image"] &&
-        [[_nativeAd.properties objectForKey:key] isKindOfClass:[NSString class]]) {
-      if ([[_nativeAd.properties objectForKey:key] isEqualToString:imageURL]) {
-        return key;
-      }
-    }
-  }
-  return nil;
-}
-
-- (void)precacheImagesWithURL:(NSArray<NSURL *> *)imageURLs {
-  id<GADMAdNetworkConnector> strongConnector = _connector;
-  _imagesDictionary = [[NSMutableDictionary alloc] init];
-
-  for (NSURL *imageURL in imageURLs) {
-    NSData *cachedImageData =
-        [[MPNativeCache sharedCache] retrieveDataForKey:imageURL.absoluteString];
-
-    UIImage *image = [UIImage imageWithData:cachedImageData];
-    if (image) {
-      // By default, the image data isn't decompressed until set on a UIImageView, on the main
-      // thread. This can result in poor scrolling performance. To fix this, we force decompression
-      // in the background before assignment to a UIImageView.
-      UIGraphicsBeginImageContext(CGSizeMake(1, 1));
-      [image drawAtPoint:CGPointZero];
-      UIGraphicsEndImageContext();
-
+- (nullable NSMutableDictionary<NSString *, GADNativeAdImage *> *)
+    imageDictionaryForImageURLDictionary:(NSDictionary<NSString *, NSURL *> *)imageURLDictionary {
+  NSMutableDictionary<NSString *, GADNativeAdImage *> *imagesDictionary =
+      [[NSMutableDictionary alloc] init];
+  for (NSString *imageKey in imageURLDictionary.allKeys) {
+    NSData *imageData = [[MPNativeCache sharedCache]
+        retrieveDataForKey:imageURLDictionary[imageKey].absoluteString];
+    if (imageData) {
+      UIImage *image = [UIImage imageWithData:imageData];
       GADNativeAdImage *nativeAdImage = [[GADNativeAdImage alloc] initWithImage:image];
-      NSString *imagekey = [self returnImageKey:imageURL.absoluteString];
-      GADMAdapterMoPubMutableDictionarySetObjectForKey(_imagesDictionary, imagekey, nativeAdImage);
+      GADMAdapterMoPubMutableDictionarySetObjectForKey(imagesDictionary, imageKey, nativeAdImage);
     }
   }
+  return imagesDictionary;
+}
 
-  if (_imagesDictionary.count < imageURLs.count) {
-    MPLogDebug(@"Cache miss on %@. Re-downloading...", imageURLs);
+- (void)preCacheNativeImagesWithCompletionHandler:
+    (void (^)(NSDictionary<NSString *, GADNativeAdImage *> *_Nullable imagesDictionary))
+        completionHandler {
+  NSDictionary<NSString *, NSURL *> *imageURLDictionary = [self imageURLs];
+  if (!imageURLDictionary[kAdMainImageKey] || !imageURLDictionary[kAdIconImageKey]) {
+    NSError *adapterError = GADMAdapterMoPubErrorWithCodeAndDescription(
+        kGADErrorReceivedInvalidResponse, @"Can't find the required MoPub native ad image assets.");
+    [_connector adapter:self didFailAd:adapterError];
+    return;
+  }
 
-    GADMAdapterMoPub __weak* weakSelf = self;
-    [_imageDownloadQueue
-        addDownloadImageURLs:imageURLs
-             completionBlock:^(NSArray *errors) {
-               GADMAdapterMoPub *strongSelf = weakSelf;
-               if (strongSelf) {
-                 if (errors.count == 0) {
-                   id<GADMAdNetworkConnector> strongConnector = strongSelf->_connector;
-                   for (NSURL *imageURL in imageURLs) {
-                     UIImage *image =
-                         [UIImage imageWithData:[[MPNativeCache sharedCache]
-                                                    retrieveDataForKey:imageURL.absoluteString]];
+  // Indicates whether the image assets should be downloaded or not.
+  BOOL shouldDownloadImages = YES;
+  if (_nativeAdOptions) {
+    for (GADAdLoaderOptions *loaderOptions in _nativeAdOptions) {
+      if ([loaderOptions isKindOfClass:[GADNativeAdImageAdLoaderOptions class]]) {
+        GADNativeAdImageAdLoaderOptions *imageOptions =
+            (GADNativeAdImageAdLoaderOptions *)loaderOptions;
+        shouldDownloadImages = !imageOptions.disableImageLoading;
+      } else if ([loaderOptions isKindOfClass:[GADNativeAdViewAdOptions class]]) {
+        _nativeAdViewAdOptions = (GADNativeAdViewAdOptions *)loaderOptions;
+      }
+    }
+  }
+  /// A dictionary that contains the icon and image assets for the native ad.
+  NSMutableDictionary<NSString *, GADNativeAdImage *> *imagesDictionary;
+  if (shouldDownloadImages) {
+    imagesDictionary = [self imageDictionaryForImageURLDictionary:imageURLDictionary];
 
-                     GADNativeAdImage *nativeAdImage =
-                         [[GADNativeAdImage alloc] initWithImage:image];
-                     NSString *imagekey = [strongSelf returnImageKey:imageURL.absoluteString];
-                     GADMAdapterMoPubMutableDictionarySetObjectForKey(strongSelf->_imagesDictionary,
-                                                                      imagekey, nativeAdImage);
-                   }
-                   if ([strongSelf->_imagesDictionary objectForKey:kAdIconImageKey] &&
-                       [strongSelf->_imagesDictionary objectForKey:kAdMainImageKey]) {
-                     strongSelf->_mediatedAd = [[MoPubAdapterMediatedNativeAd alloc]
-                                              initWithMoPubNativeAd:strongSelf->_nativeAd
-                                  mappedImages:strongSelf->_imagesDictionary
-                           nativeAdViewOptions:strongSelf->_nativeAdViewAdOptions
-                                 networkExtras:[strongConnector networkExtras]];
-                     [strongConnector adapter:strongSelf
-                         didReceiveMediatedUnifiedNativeAd:strongSelf->_mediatedAd];
-                   }
-                 } else {
-                   MPLogDebug(@"Failed to download images. Giving up for now.");
-                   NSError *adapterError = [NSError errorWithDomain:kGADMAdapterMoPubErrorDomain
-                                                               code:kGADErrorNetworkError
-                                                           userInfo:nil];
-                   [strongConnector adapter:strongSelf didFailAd:adapterError];
+    // Check if MoPub image assets are cached.
+    if (!imagesDictionary[kAdIconImageKey] || !imagesDictionary[kAdMainImageKey]) {
+      GADMAdapterMoPub __weak *weakSelf = self;
+      [_imageDownloadQueue
+          addDownloadImageURLs:imageURLDictionary.allValues
+               completionBlock:^(NSArray *errors) {
+                 GADMAdapterMoPub *strongSelf = weakSelf;
+                 if (!strongSelf) {
+                   NSLog(@"MPNativeAd deallocated before MoPub native ad images were downloaded.");
                    return;
                  }
-               } else {
-                 MPLogDebug(
-                     @"MPNativeAd deallocated before loadImageForURL:intoImageView: download "
-                     @"completion block was called");
-                 NSError *adapterError = [NSError errorWithDomain:kGADMAdapterMoPubErrorDomain
-                                                             code:kGADErrorInternalError
-                                                         userInfo:nil];
-                 [strongConnector adapter:strongSelf didFailAd:adapterError];
-                 return;
-               }
-             }];
+                 NSDictionary<NSString *, GADNativeAdImage *> *imagesDictionary =
+                     [strongSelf imageDictionaryForImageURLDictionary:imageURLDictionary];
+                 completionHandler(imagesDictionary);
+               }];
+      return;
+    }
   } else {
-    if (_shouldDownloadImages) {
-      _mediatedAd = [[MoPubAdapterMediatedNativeAd alloc]
-          initWithMoPubNativeAd:_nativeAd
-                   mappedImages:_imagesDictionary
-            nativeAdViewOptions:_nativeAdViewAdOptions
-                  networkExtras:[strongConnector networkExtras]];
-      [strongConnector adapter:self didReceiveMediatedUnifiedNativeAd:_mediatedAd];
-    } else {
-      NSMutableDictionary *_mainImageDictionary = [[NSMutableDictionary alloc] init];
-      GADNativeAdImage *_tempMainImage = _imagesDictionary[kAdMainImageKey];
-      _mainImageDictionary[kAdMainImageKey] = _tempMainImage;
-
-      _mediatedAd = [[MoPubAdapterMediatedNativeAd alloc]
-          initWithMoPubNativeAd:_nativeAd
-                   mappedImages:_mainImageDictionary
-            nativeAdViewOptions:_nativeAdViewAdOptions
-                  networkExtras:[strongConnector networkExtras]];
-      [strongConnector adapter:self didReceiveMediatedUnifiedNativeAd:_mediatedAd];
+    for (NSString *imageKey in imageURLDictionary.allKeys) {
+      GADNativeAdImage *nativeAdImage =
+          [[GADNativeAdImage alloc] initWithURL:imageURLDictionary[imageKey] scale:1.0];
+      GADMAdapterMoPubMutableDictionarySetObjectForKey(imagesDictionary, imageKey, nativeAdImage);
     }
   }
+  completionHandler(imagesDictionary);
 }
 
 #pragma mark MPNativeAdDelegate Methods