+++ adapters/MoPub/CHANGELOG.md
@@ -4,7 +4,7 @@
 - Verified compatibility with MoPub SDK 5.8.0.
 - Now requires minimum iOS version of 9.0.
 - Removed support for Native App Install ad requests. Apps must use the Unified Native Ads API to request native ads.
-- Now requires Google Mobile Ads SDK version 7.46.0 or higher.
+- Now requires Google Mobile Ads SDK version 7.51.0 or higher.
 
 #### Version 5.7.1.0
 - Verified compatibility with MoPub SDK 5.7.1.